Django-Middleware-0x03/requirements.txt
asgiref==3.9.1
Django==4.2.24
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0
